Speech/requirements.txt
absl-py==2.2.2
accelerate==1.6.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
angles==2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens
attrdict==2.0.1
attrs==25.3.0
audioread==3.0.1
binpacking==1.5.2
bitsandbytes==0.45.5
bond==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm
contourpy==1.3.2
crcmod==1.7
cryptography==44.0.3
cycler==0.12.1
dacite==1.9.2
datasets==3.5.1
debugpy
decorator
deepspeed==0.16.3
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
editdistance==0.8.1
einops==0.8.1
exceptiongroup
executing
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
funasr==1.2.6
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
grpcio==1.71.0
gxl_ai_utils==1.5.2
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
jaconv==0.4.0
jamo==0.4.1
jedi
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.5.0
jsonlines==4.0.0
jupyter_client
jupyter_core
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.7.0
matplotlib==3.10.1
matplotlib-inline
mdurl==0.1.2
modelscope==1.19.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
nest_asyncio
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.77.0
openai-whisper==20250625
orjson==3.10.18
oss2==2.19.1
packaging
pandas==2.2.3
parso
peft==0.14.0
pexpect
pickleshare
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
prompt_toolkit
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess
pure_eval
py-cpuinfo==9.0.0
pyahocorasick==2.2.0
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil
python-multipart==0.0.20
pytorch-wpe==0.0.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq
regex==2024.11.6
requests==2.32.3
# resource-retriever==3.1.2
rich==14.0.0
ruff==0.11.8
safehttpx==0.1.6
safetensors==0.5.3
samplerate==0.2.1
scikit-learn==1.6.1
scipy==1.15.2
# scripts==3.7.0
scripts==3.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==69.5.1
shellingham==1.5.4
simplejson==3.20.1
six
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
stack_data
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
# torch==2.6.0
torch-complex==0.4.4
# torchaudio==2.6.0
# torchvision==0.21.0
tornado
tqdm==4.67.1
traitlets
transformers==4.49.0
transformers-stream-generator==0.0.5
triton==3.2.0
trl==0.15.2
typeguard==4.4.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.19.8
wcwidth
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0